Add TaskScheduler to handle the ComRuntime Update the README, CHANGELOG Remove build.log and add it to .gitignore Update time_task example with changes (rest to come later)
