+++ README.md
@@ -54,5 +54,29 @@ All settings are available for the tasks.
 The documentation contains all relevant information from the
 [Microsoft Task Scheduler documentation](https://learn.microsoft.com/en-us/windows/win32/taskschd/task-scheduler-reference).
 
+## Running examples
+After cloning the repository, examples can be run using either 
+
+`cargo run --example <name>` to run a specific example or,
+`cargo run --examples` to run all examples.
+
 ## Changelog
 See the [changelog file](CHANGELOG.md).
+
+## Upgrade Guide (0.* to 1.0)
+
+### ComRuntime
+
+#### Creating
+The `ComRuntime` is now handled by the `TaskScheduler` and should be created using: 
+```rust
+    let ts = TaskScheduler::new()?;
+    let com = ts.get_com();
+    let sb = ScheduleBuilder::new(&com).unwrap();
+    // ... snip
+```
+
+#### Uninilizing
+`ScheduleBuilder`'s no longer need to be manually uninitialized. In pre-1.0, `ScheduleBuilding::uninitalize()` would call `CoUninitialize` which would effectively close the COM. This could be problamatic if you had multiple builders or schedules built since having n+1 COMs would simply reuse the initial COM.
+
+The changes brought in 1.0 will now keep an `Rc<Com>` privately stored in the `TaskScheduler`. Now when all the references are dropped, the COM uninitialization will be handled "automagically".
